# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: soukMA - Moroccan e-commerce marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me/become-vendor:
    post:
      tags: [Me]
      operationId: becomeVendor
      summary: Promote myself to vendor (provides shop info)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BecomeVendorBody'
      responses:
        '200':
          description: Vendor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfile'
...
  /cart/items:
    post:
      tags: [Cart]
      operationId: addCartItem
      summary: Add a product to the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddCartItemBody'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
  /cart/items/{itemId}:
    patch:
      tags: [Cart]
      operationId: updateCartItem
      summary: Update a cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemBody'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
    delete:
      tags: [Cart]
      operationId: removeCartItem
      summary: Remove a cart item
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
...
  /orders:
    get:
      tags: [Orders]
      operationId: listMyOrders
      summary: List my orders (customer)
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags: [Orders]
      operationId: placeOrder
      summary: Convert cart to an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderBody'
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
  /orders/{id}:
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Get an order by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
  /orders/{id}/status:
    patch:
      tags: [Orders]
      operationId: updateOrderStatus
      summary: Update order status (vendor or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusBody'
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
...
components:
...
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id: { type: string }
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
...
    MyProfile:
      type: object
      required: [authenticated, role]
      properties:
        authenticated: { type: boolean }
        role:
          type: string
          enum: [guest, customer, vendor, admin]
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
        vendor:
          oneOf:
            - $ref: '#/components/schemas/Vendor'
            - type: 'null'
    BecomeVendorBody:
      type: object
      required: [shopName, phone, city]
      properties:
        shopName: { type: string, minLength: 2, maxLength: 80 }
        description: { type: string, maxLength: 500 }
        phone: { type: string, minLength: 6, maxLength: 30 }
        city: { type: string, minLength: 2, maxLength: 80 }
    Vendor:
      type: object
      required: [id, shopName, city, phone, createdAt]
      properties:
        id: { type: string }
        shopName: { type: string }
        description:
          type: [string, 'null']
        phone: { type: string }
        city: { type: string }
        createdAt: { type: string, format: date-time }
...
    Cart:
      type: object
      required: [items, subtotal, currency, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal: { type: number }
        currency: { type: string }
        itemCount: { type: integer }
    CartItem:
      type: object
      required: [id, productId, productTitle, productImage, unitPrice, quantity, lineTotal]
      properties:
        id: { type: string }
        productId: { type: string }
        productTitle: { type: string }
        productImage:
          type: [string, 'null']
        unitPrice: { type: number }
        quantity: { type: integer }
        lineTotal: { type: number }
    AddCartItemBody:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1, maximum: 99 }
    UpdateCartItemBody:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1, maximum: 99 }
    Order:
      type: object
      required: [id, reference, status, total, currency, itemCount, paymentMethod, createdAt]
      properties:
        id: { type: string }
        reference: { type: string }
        status:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
        total: { type: number }
        currency: { type: string }
        itemCount: { type: integer }
        paymentMethod:
          type: string
          enum: [cmi, cih_pay, cash_on_delivery]
        createdAt: { type: string, format: date-time }
    OrderItem:
      type: object
      required: [id, productId, productTitle, productImage, unitPrice, quantity, lineTotal, vendorName]
      properties:
        id: { type: string }
        productId: { type: string }
        productTitle: { type: string }
        productImage:
          type: [string, 'null']
        unitPrice: { type: number }
        quantity: { type: integer }
        lineTotal: { type: number }
        vendorName: { type: string }
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/Order'
        - type: object
          required: [items, shipping]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            shipping:
              $ref: '#/components/schemas/ShippingAddress'
    ShippingAddress:
      type: object
      required: [fullName, phone, address, city]
      properties:
        fullName: { type: string }
        phone: { type: string }
        address: { type: string }
        city: { type: string }
        notes:
          type: [string, 'null']
    PlaceOrderBody:
      type: object
      required: [shipping, paymentMethod]
      properties:
        shipping:
          $ref: '#/components/schemas/ShippingAddress'
        paymentMethod:
          type: string
          enum: [cmi, cih_pay, cash_on_delivery]
    UpdateOrderStatusBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, paid, processing, shipped, delivered, cancelled]
...