# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * soukMA - Moroccan e-commerce marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  reference: string;
  status: OrderStatus;
  total: number;
  currency: string;
  itemCount: number;
  paymentMethod: OrderPaymentMethod;
  createdAt: Date;
}